+++ src/main.py
@@ -4,9 +4,11 @@ from fastapi.templating import Jinja2Templates
 from fastapi.staticfiles import StaticFiles  # <--- New Import
 from datetime import datetime
 from src.config import settings
+from src.router import api_router
 from src.__init__ import __version__ as version
 
 app = FastAPI(title=settings.APP_NAME, __version__ = version)
+app.include_router(api_router)
 
 # Mount Static Files
 # This tells FastAPI: "Any request starting with /static should look in src/static"