+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.1] - 2026-05-05
 
 ### Added
+- health endpoint
+- test for health endpoint
 
 ### Changed
 - Adding cleanup commands for EC2 instance in `deploy.yml` file